refactor: создать мапу, раскидать компоненты по папкам
