+++ src/components/burger-ingredients/burger-ingredients.module.css
@@ -5,16 +5,6 @@
   flex-direction: column;
 }
 
-.nonList {
-  list-style: none;
-  margin: 0;
-  padding: 0;
-}
-
-.nonLink {
-  text-decoration: none;
-}
-
 .tab {
   display: flex;
   margin-bottom: 40px;
