+++ src/components/card-order/card-order.module.css
@@ -1,0 +1,40 @@
+.orderComponentsAndPrice {
+  display: flex;
+  flex-direction: row;
+  width: 100%;
+  align-items: center;
+  justify-content: space-between;
+
+}
+
+.orderComponents {
+  display: flex;
+  flex-direction: row;
+
+  width: 304px;
+  height: 64px;
+}
+
+.orderId {
+  display: flex;
+  justify-content: space-between;
+  flex-direction: row;
+}
+
+.orderPrice {
+  display: flex;
+  flex-direction: row;
+}
+
+.cardOrder {
+  width: 100%;
+  max-width: calc(100% - 8px);
+  box-sizing: border-box;
+  border-radius: 40px;
+  padding: 24px;
+  display: flex;
+  flex-direction: column;
+  gap: 24px;
+  background: #1c1c21;
+  box-shadow: 0 0 1px 0 rgba(0, 0, 0, 0.04), 0 0 2px 0 rgba(0, 0, 0, 0.06), 0 4px 8px 0 rgba(0, 0, 0, 0.04);
+}
