+++ src/components/feed-orders-profile/feed-orders-profile.module.css
@@ -1,58 +1,3 @@
-
-.cardOrder {
-  width: 100%;
-  max-width: calc(100% - 8px);
-  box-sizing: border-box;
-  border-radius: 40px;
-  padding: 24px;
-  display: flex;
-  flex-direction: column;
-  gap: 24px;
-  background: #1c1c21;
-  box-shadow: 0 0 1px 0 rgba(0, 0, 0, 0.04), 0 0 2px 0 rgba(0, 0, 0, 0.06), 0 4px 8px 0 rgba(0, 0, 0, 0.04);
-}
-
-.cardOrder2 {
-  width: 100%;
-  max-width: 640px;
-  box-sizing: border-box;
-  display: flex;
-  flex-direction: column;
-  /*justify-content: center;*/
-  /*  align-items: center;*/
-  margin: 0 auto;
-  padding-top: 82px;
-
-}
-
-.cardOrder3 {
-  width: 100%;
-  max-width: 640px;
-  box-sizing: border-box;
-  display: flex;
-  flex-direction: column;
-  /*justify-content: center;*/
-  /*  align-items: center;*/
-  margin: 0 auto;
-  padding-top: 20px;
-
-}
-
-
-.orderId {
-  display: flex;
-  justify-content: space-between;
-  flex-direction: row;
-}
-
-.orderComponents {
-  display: flex;
-  flex-direction: row;
-
-  width: 304px;
-  height: 64px;
-}
-
 .containerFeed {
   display: flex;
   height: 100%;
@@ -62,55 +7,11 @@
   overflow-y: scroll;
 }
 
-.containerFeed2 {
-  display: flex;
-  max-height: 312px;
-  width: 100%;
-  flex-direction: column;
-  gap: 16px;
-  overflow-y: scroll;
-
-}
-
-
-.rowIng {
-  display: flex;
-  flex-direction: row;
-  gap: 16px;
-  justify-content: space-between;
-  align-items: center;
-  padding-right: 24px;
-}
-
-
-.orderComponentsAndPrice {
-  display: flex;
-  flex-direction: row;
-  width: 100%;
-  align-items: center;
-  justify-content: space-between;
-
-}
-
-.orderPrice {
-  display: flex;
-  flex-direction: row;
-}
-
-.orderPrice2 {
-  display: flex;
-  flex-direction: row;
-  margin-left: auto;
-}
-
-
 .imgCardSmall {
   height: 60px;
   width: 60px;
-
   border-radius: 100px;
   box-sizing: border-box;
-  /*  padding: 10px;*/
   background: black 50% / 180% 100% no-repeat;
 }
 
@@ -126,65 +27,4 @@
   height: 64px;
   padding: 0;
   border-radius: 100px;
-  /*  position: absolute;*/
 }
-
-.absolute {
-  position: absolute;
-}
-
-.relative {
-  position: relative;
-}
-
-.imgCardSmallFonNumber {
-  height: 100%;
-  width: 100%;
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-  border-radius: 100px;
-  box-sizing: border-box;
-  /*  padding: 10px;*/
-  background: rgba(0, 0, 0, 0.6) 50% / 180% 100% no-repeat;
-}
-
-.rowIngNew {
-  width: 64px;
-  height: 64px;
-  position: relative;
-  margin: 10px;
-  box-sizing: border-box;
-  background: black 50% / 180% 100% no-repeat;
-  border-radius: 50%;
-}
-
-
-.rowIngNew:after {
-  width: 68px;
-  height: 68px;
-  content: "";
-  position: absolute;
-  box-sizing: border-box;
-  background-image: linear-gradient(63deg, #801AB2 0%, #4C4CFF 100%);
-  border-radius: 50%;
-  z-index: -4;
-  left: -2px;
-  top: -2px
-}
-
-/*
-.buttonCardSmall {
-  overflow: hidden;
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-  border: 2px solid inherit;
-  box-sizing: border-box;
-
-  padding: 0;
-  border-radius: 100px;
-  !*  position: absolute;*!
-}*/
