+++ src/pages/feed.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import {FeedHistory} from "../components/feed-orders-profile/feed-orders-profile";
+import {Orders} from "../components/feed-orders-profile/feed-orders-profile";
 import {Stats} from "../components/stats/stats";
 import {useDispatch, useSelector} from "react-redux";
 import {Modal} from "../components/modal/modal";
@@ -20,7 +20,7 @@ export function Feed() {
     <>
       <section className={'pl-5 pr-5 half-home'}>
         <p className="text text_type_main-large pb-5">Лента заказов</p>
-        <FeedHistory/>
+        <Orders/>
       </section>
       <section className={'pl-5 pr-5 pt-15 half-home'}>
         <Stats/>
