--- src/pages/orders.jsx
@@ -1,17 +1,0 @@
-import React from "react";
-import {FeedHistory} from "../components/feed-orders-profile/feed-orders-profile";
-
-//Заглушка
-export function Orders() {
-
-  return (
-<div className='feedHistoryOrders'
->
-  <FeedHistory/>
-
-
-</div>
-
-
-  )
-}
