+++ src/pages/profile-orders.jsx
@@ -1,0 +1,10 @@
+import React from "react";
+import {Orders} from "../components/feed-orders-profile/feed-orders-profile";
+
+
+export function ProfileOrders() {
+
+  return (
+    <Orders/>
+  )
+}
